+++ src/gc.rs
@@ -1,3 +1,5 @@
+use crate::Janet;
+
 // TODO: Should these types be Send and/or Sync?
 
 /// The Janet Garbage Collector type.
@@ -52,6 +54,25 @@ impl JanetGc {
     pub fn unlock(guard: JanetGcLockGuard) {
         drop(guard)
     }
+
+    /// Roots the `value` to the GC. This prevents the GC from removing the `value` and
+    /// all of its children in a garbage collection.
+    ///
+    /// The returned guard will [`unroot`](JanetGc::unroot) the `value` when dropped.
+    #[inline]
+    pub fn root(&self, value: Janet) -> JanetGcRootGuard {
+        JanetGcRootGuard::new(value)
+    }
+
+    /// Immediately drops the guard, and consequently unlocks the Janet GC.
+    ///
+    /// This function is equivalent to calling [`drop`] on the guard but is more
+    /// self-documenting. Alternately, the guard will be automatically dropped when it
+    /// goes out of scope.
+    #[inline]
+    pub fn unroot(guard: JanetGcRootGuard) {
+        drop(guard)
+    }
 }
 
 /// An RAII implementation of a “scoped lock” of the Janet GC. When this structure is
@@ -62,6 +83,7 @@ pub struct JanetGcLockGuard {
 }
 
 impl JanetGcLockGuard {
+    #[inline]
     pub(crate) fn new(handle: i32) -> Self {
         Self { handle }
     }
@@ -71,5 +93,32 @@ impl Drop for JanetGcLockGuard {
     #[inline]
     fn drop(&mut self) {
         unsafe { evil_janet::janet_gcunlock(self.handle) }
+    }
+}
+
+/// An RAII implementation of a rooted [`Janet`] value. When this structure is dropped
+/// (falls out of scope), the rooting will be undone.
+#[derive(Debug)]
+pub struct JanetGcRootGuard {
+    value: Janet,
+}
+
+impl JanetGcRootGuard {
+    #[inline]
+    fn new(value: Janet) -> Self {
+        unsafe { evil_janet::janet_gcroot(value.inner) };
+        Self { value }
+    }
+}
+
+impl Drop for JanetGcRootGuard {
+    #[inline]
+    fn drop(&mut self) {
+        // SAFETY: Since we unrooting the same value we rooted, this should always work.
+        // For extra safety, below it's add a debug assert to be sure on debug compilations.
+        let res = unsafe { evil_janet::janet_gcunroot(self.value.inner) };
+
+        // Assert in debug mode that the result is 1
+        debug_assert_eq!(res, 1)
     }
 }