+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to the library should be put here
 -   Add `dedup`, `dedup_by` and `dedup_by_key` for `JanetArray`
 -   Add `get_method` and `has_method` to `Janet`
 -   Add `JanetGc` and `JanetGcLockGuard` types to access some Janet GC operations
+-   Add `JanetGcRootGuard` and the functions `JanetGc::root` and `JanetGc::unroot` to root a Janet object to the GC
 -   Add functions to get reference to a `JanetAbstract` data safely
 -   Add `JanetAbstract::is`
 -   Add `Janet::int64`